+++ packages/api/Tsz.Api/Program.cs
@@ -16,7 +16,6 @@ using Tsz.Modules.LeaveTypes;
 using Tsz.Modules.Timesheets;
 using Tsz.Modules.Users;
 using Tsz.Modules.Users.Seeding;
-using Tsz.Modules.Workdays;
 
 JwtSecurityTokenHandler.DefaultMapInboundClaims = false;
 
@@ -47,7 +46,7 @@ builder.Services.AddSingleton(TimeProvider.System);
 builder.Services.AddHttpContextAccessor();
 builder.Services.AddScoped<ICurrentUser, HttpContextCurrentUser>();
 
-IReadOnlyList<IModule> modules = [new UsersModule(), new CustomersModule(), new ContractsModule(), new LeaveTypesModule(), new WorkdaysModule(), new TimesheetsModule()];
+IReadOnlyList<IModule> modules = [new UsersModule(), new CustomersModule(), new ContractsModule(), new LeaveTypesModule(), new TimesheetsModule()];
 foreach (var m in modules) m.RegisterServices(builder.Services, builder.Configuration);
 
 var app = builder.Build();